requirements.txt
django==5.0.6 
djangorestframework==3.15.1
django-bootstrap-v5==1.0.11
pandas==2.0.1
openpyxl==3.1.3
